requirements.txt
aerich==0.7.1
aiofiles==22.1.0
aiogram==3.0.0b6
aiohttp==3.8.3
aioredis==2.0.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.6.2
APScheduler==3.9.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
Babel==2.11.0
bcrypt==4.0.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
dictdiffer==0.9.0
facebook-sdk==3.1.0
fastapi==0.85.2
fastapi-admin==1.0.4
fastapi-utils==0.2.1
frozenlist==1.3.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
magic-filter==1.0.9
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
pendulum==2.1.2
pydantic==1.10.2
pyparsing==3.0.9
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.43
starlette==0.20.4
tomlkit==0.11.5
tortoise-orm==0.19.2
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.19.0
watchfiles==0.18.0
websockets==10.3
wrapt==1.14.1
yarl==1.8.1
